requirements.txt
numpy>=1.21.0
pandas>=1.3.0
yfinance>=0.1.87
tensorflow==2.18.0
scikit-learn>=1.0.0
streamlit>=1.28.0
matplotlib>=3.5.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=1.0.0
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=3.0.0
faiss-cpu>=1.7.4